pyproject.toml
[project]
name = "repo-for-dummies"
version = "0.1.0"
description = "AI agent that converts GitHub repos into beginner-friendly guides"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.11",
    "langchain-mistralai>=1.1.2",
    "langchain[google-genai]>=1.2.12",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]